refactor(agents): share composer-input slash grammar via runtime
Lift the slash-command serializer and grammar out of the DOM-only
ComposerEditor into agents-runtime/src/composer-input.ts so every composer
surface (desktop ProseMirror, upcoming mobile native TextInput) derives
nodes, highlighting, and the autocomplete trigger from one source of truth.

- Move serializeComposerInput, normalizeCommandName, and
  formatSlashCommandArgumentHint into composer-input.ts; export via /client.
- Extract both slash regexes as named constants: createSlashCommandTokenRegex
  (token) and SLASH_COMMAND_TRIGGER_REGEX (trigger), plus a reusable
  detectSlashCommandTrigger helper for trigger detection on plain strings.
- Repoint ComposerEditor's decoration/getSlashQuery regexes and the
  MessageInput/NewSessionView call sites at the shared exports.
- Move the serializer/hint tests into the runtime suite and add
  detectSlashCommandTrigger coverage; the dropped-token test pins the v1
  grammar boundary shared by desktop fallback and mobile.
