+++ packages/agents-runtime/src/client.ts
@@ -26,7 +26,15 @@ export {
   manifestAttachmentKey,
 } from './manifest-helpers'
 export { buildSections, buildTimelineEntries } from './use-chat'
-export { COMPOSER_INPUT_MESSAGE_TYPE } from './composer-input'
+export {
+  COMPOSER_INPUT_MESSAGE_TYPE,
+  serializeComposerInput,
+  normalizeCommandName,
+  formatSlashCommandArgumentHint,
+  detectSlashCommandTrigger,
+  createSlashCommandTokenRegex,
+  SLASH_COMMAND_TRIGGER_REGEX,
+} from './composer-input'
 
 export type {
   EntityStreamDB,
@@ -37,6 +45,7 @@ export type {
   ComposerInputPayload,
   ComposerNode,
   SlashCommandRow,
+  SlashCommandTrigger,
 } from './composer-input'
 export type { AgentsClient, AgentsClientConfig } from './agents-client'
 export type {
